+++ 04.class/memo.js
@@ -1,16 +1,123 @@
-import MemoApp from "./memo_app.js";
+import sqlite3 from "sqlite3";
+import enquirer from "enquirer";
+import readline from "readline";
 
-const memo = async function () {
-  try {
-    const memoapp = await new MemoApp(process.argv[2]);
-    await memoapp.optionControllers(memoapp);
-  } catch (error) {
-    if (error instanceof Error) {
-      console.error(error.message);
-    } else {
-      console.error(new Error(error));
-    }
+class Memo {
+  constructor(filename) {
+    this.db = new sqlite3.Database(filename);
+    this.createDatabase;
   }
-};
 
-memo();
+  createDatabase() {
+    return new Promise((resolve, reject) => {
+      this.db.run(
+        "CREATE TABLE IF NOT EXISTS memos (id INTEGER PRIMARY KEY AUTOINCREMENT, title TEXT, context TEXT)",
+        (error) => {
+          if (error) {
+            reject(error);
+          } else {
+            resolve();
+          }
+        },
+      );
+    });
+  }
+
+  add() {
+    return new Promise((resolve, reject) => {
+      try {
+        const tmpMemoList = [];
+        const rl = readline.createInterface({
+          input: process.stdin,
+        });
+
+        rl.on("line", (memoData) => {
+          tmpMemoList.push(memoData);
+        });
+
+        rl.on("close", () => {
+          const title = tmpMemoList[0];
+          const context = tmpMemoList.join("\n");
+          this.db.run(
+            "INSERT INTO memos (title, context) VALUES (?, ?)",
+            [title, context],
+            function () {},
+          );
+          resolve();
+        });
+      } catch (error) {
+        reject(error);
+      }
+    });
+  }
+
+  async list() {
+    return new Promise((resolve, reject) => {
+      this.db.all("SELECT * FROM memos ORDER BY id ASC", (error, memos) => {
+        if (error) {
+          reject(error);
+        } else {
+          resolve(memos);
+        }
+      });
+    });
+  }
+
+  async delete(deleteMessage) {
+    const memoId = await this.getChoiceMemoId(deleteMessage);
+    return await new Promise((resolve) => {
+      this.db.run("DELETE FROM memos WHERE id = ?", memoId, () => {
+        resolve(memoId);
+      });
+    });
+  }
+
+  async reference(referenceMessage) {
+    const memoId = await this.getChoiceMemoId(referenceMessage);
+    return await new Promise((resolve) => {
+      this.db.get(
+        "SELECT context FROM memos WHERE id = ?",
+        memoId,
+        (error, searchResultMemo) => {
+          if (error) {
+            console.error(error.message);
+          } else {
+            resolve(searchResultMemo);
+          }
+        },
+      );
+    });
+  }
+
+  async getChoiceMemoId(Message) {
+    const choices = await new Promise((resolve) => {
+      this.db.all("SELECT * FROM memos", (error, memos) => {
+        if (error) {
+          console.error(error.message);
+        } else {
+          resolve(
+            memos.map((memo) => ({
+              name: memo.title,
+              value: memo.id,
+            })),
+          );
+        }
+      });
+    });
+
+    const question = {
+      type: "select",
+      message: `${Message}`,
+      name: "memoId",
+      choices: choices,
+      result() {
+        return this.focused.value;
+      },
+    };
+
+    const answer = await enquirer.prompt(question);
+    return answer.memoId;
+  }
+}
+
+export default Memo;
